metrics_collector/requirements.txt
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
prometheus-client==0.21.1
psycopg2-binary==2.9.10
requests==2.32.3
six==1.17.0
urllib3==2.2.3
